pyproject.toml
[project]
name = "homecook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-playwright>=0.7.2",
    "tomlkit>=0.14.0",
]
